Change non-blocking tasks' behavior - detach them from the parent pipeline.
Blocking tasks are attached to the current pipeline, which means that
they will be cancelled if the parent pipeline is cancelled. This behavior
was previously present on the non-blocking tasks too.

This basically made non-blocking tasks unusable beacause exiting from
the pipeline (when all blocks are evaluated) would cancel them too.
Now the non-blocking tasks are detached from the parent pipeline and will
continue running even if the pipeline is cancelled or exited. Non-blocking
tasks MUST be explicitely cancelled or stopped with the cancel-task!
or stop-task! functions.